Moved "Device Graph" page out of the main views and into a popup and the menu
