+++ src/lib/y2partitioner/widgets/pages.rb
@@ -42,4 +42,3 @@ require "y2partitioner/widgets/pages/bcache.rb"
 require "y2partitioner/widgets/pages/bcaches.rb"
 require "y2partitioner/widgets/pages/btrfs_filesystems.rb"
 require "y2partitioner/widgets/pages/btrfs.rb"
-require "y2partitioner/widgets/pages/device_graph.rb"
